requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.30
openai==1.35.6
anthropic==1.17.0
google-generativeai==0.5.4
gunicorn==22.0.0
